schedule/requirements.txt
aniso8601==9.0.1
autopep8==2.0.2
blinker==1.6.2
cachecontrol==0.13.1
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cryptography==41.0.1
firebase-admin==6.2.0
Flask==2.3.2
Flask-Cors==4.0.0
Flask-Jsonpify==1.5.0
Flask-RESTful==0.3.10
flask-sqlalchemy==3.0.5
google-api-core==2.11.1
google-api-python-client==2.90.0
google-auth==2.21.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.11.1
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.1
greenlet==2.0.2
grpcio==1.56.0
grpcio-status==1.56.0
httplib2==0.22.0
idna==3.4
importlib-metadata==6.7.0
instagrapi==1.17.13
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
msgpack==1.0.5
Pillow==10.0.0
proto-plus==1.22.3
protobuf==4.23.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.10.0
pycparser==2.21
pycryptodomex==3.18.0
pydantic==1.10.9
PyJWT==2.7.0
pyparsing==3.1.0
PySocks==1.7.1
pytz==2023.3
requests==2.31.0
rsa==4.9
six==1.16.0
SQLAlchemy==2.0.17
tomli==2.0.1
typing-extensions==4.6.3
uritemplate==4.1.1
urllib3==1.26.16
Werkzeug==2.3.6
zipp==3.15.0
